requirements.txt
openai==0.28.1
faiss-cpu==1.7.4
numpy==1.24.3
pandas==1.5.3
gradio==3.50.2
python-dotenv==1.0.0
requests==2.31.0
httpx==0.24.1